+++ Branch-Xamarin-SDK/Requests/BranchOpenRequest.cs
@@ -22,6 +22,7 @@ namespace BranchXamarinSDK
 			public string push_identifier;
 			public string external_intent_uri;
 			public string external_intent_extra;
+			public string universal_link_url;
 
 			public OpenParams() {
 			}
@@ -50,6 +51,7 @@ namespace BranchXamarinSDK
 			LocalParams.push_identifier = Branch.GetInstance().PushIdentifier;
 			LocalParams.external_intent_uri = Branch.GetInstance().ExternalUri;
 			LocalParams.external_intent_extra = Branch.GetInstance().ExternalExtra;
+			LocalParams.universal_link_url = Branch.GetInstance ().UniversalLink;
 
 			Callback = callback;
 		}