+++ ChangeLog.md
@@ -4,7 +4,8 @@ Branch Xamarin SDK change log
   * BranchAndroid.GetInstance() has been renamed BranchAndroid.getInstance().
   * Automatical session management has been added. Use AndroidBranch.SetLifeCycleHandlerCallback(Context, IBranchSessionInterface) to enable automatical sesiion management. Use Branch.AutoSessionEnabled to check what type of session management is used.
   * AndroidAppLink support added.
-  * Push Notification support added.
+  * Android Push Notification support added.
+  * iOS Universal Links suppor added.
 
 - v1.1.0
   * Usage of app key is now deprecated. Use your Branch key, found in the settings page of your Dashboard if possible.